# sources/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/item-characteristics": {
      "post": {
        "operationId": "ItemCharacteristicsController_create",
        "summary": "Создание записи",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemCharacteristicDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Запись создана успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-characteristics",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "ItemCharacteristicsController_findAll",
        "summary": "Получить список записей",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Получили список записей",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemCharacteristicWithIdDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-characteristics",
          "any"
        ]
      }
    },
    "/api/v1/item-characteristics/bulk": {
      "post": {
        "operationId": "ItemCharacteristicsController_createBulk",
        "summary": "Создание несколько записей за раз",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBulkItemCharacteristicDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Записи добавлены в БД",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-characteristics",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "ItemCharacteristicsController_updateBulk",
        "summary": "Обновление записи по id",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBulkItemCharacteristicDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-characteristics",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/item-characteristics/{id}": {
      "get": {
        "operationId": "ItemCharacteristicsController_findOne",
        "summary": "Получить запись по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Получили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemCharacteristicWithIdDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-characteristics",
          "any"
        ]
      },
      "patch": {
        "operationId": "ItemCharacteristicsController_update",
        "summary": "Обновление записи по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemCharacteristicDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-characteristics",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ItemCharacteristicsController_remove",
        "summary": "Удаление записи по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-characteristics",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/v1/items/bulk": {
      "post": {
        "operationId": "ItemsController_createBulk",
        "summary": "Создание несколько записей за раз транзакцией",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBulkItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Записи добавлены в БД. Транзакция выполнена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзакция не выполнена.Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "ItemsController_updateBulk",
        "summary": "Обновление записи по uuid транзакцией",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBulkItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили запись по uuid. Транзакция выполнена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзакция не выполнена.Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/v1/items/{id}": {
      "get": {
        "operationId": "ItemsController_findOne",
        "summary": "Получить запись по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Получили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEntity"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "any"
        ]
      },
      "patch": {
        "operationId": "ItemsController_update",
        "summary": "Обновление записи по uuid транзакцией",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили запись по uuid. Транзакция выполнена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзакция не выполнена.Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ItemsController_remove",
        "summary": "Удаление записи по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "REST API документация",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "access token - токен доступа",
        "in": "header"
      },
      "refresh-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "refresh token - токен обновления",
        "in": "header"
      }
    },
    "schemas": {
...
      "HttpResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "сообщение"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CreateItemCharacteristicDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "В коробке"
          }
        },
        "required": [
          "dp_name"
        ]
      },
      "ItemCharacteristicNoIdDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "В коробке"
          }
        },
        "required": [
          "dp_name"
        ]
      },
      "CreateBulkItemCharacteristicDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemCharacteristicNoIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
      "ItemCharacteristicWithIdDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "В коробке"
          },
          "dp_id": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "dp_name",
          "dp_id"
        ]
      },
      "UpdateItemCharacteristicDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "В коробке"
          }
        },
        "required": [
          "dp_name"
        ]
      },
      "UpdateBulkItemCharacteristicDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemCharacteristicWithIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
...
      "CreateBulkItemDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemNoIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
...
      "LstItemCharacteristicEntity": {
        "type": "object",
        "properties": {
          "dp_id": {
            "type": "number"
          },
          "dp_itemId": {
            "type": "string"
          },
          "dp_characteristicId": {
            "type": "number"
          },
          "dp_value": {
            "type": "string"
          }
        },
        "required": [
          "dp_id",
          "dp_itemId",
          "dp_characteristicId",
          "dp_value"
        ]
      },
      "LstItemGaleryEntity": {
        "type": "object",
        "properties": {
          "dp_id": {
            "type": "number"
          },
          "dp_itemId": {
            "type": "string"
          },
          "dp_photoUrl": {
            "type": "string",
            "example": "https://example.com/image.png"
          }
        },
        "required": [
          "dp_id",
          "dp_itemId",
          "dp_photoUrl"
        ]
      },
      "ItemEntity": {
        "type": "object",
        "properties": {
          "dp_id": {
            "type": "string"
          },
          "dp_name": {
            "type": "string"
          },
          "dp_model": {
            "type": "string"
          },
          "dp_cost": {
            "type": "number"
          },
          "dp_photoUrl": {
            "type": "string"
          },
          "dp_itemCategoryId": {
            "type": "number"
          },
          "dp_seoKeywords": {
            "type": "string"
          },
          "dp_seoDescription": {
            "type": "string"
          },
          "dp_itemCharecteristics": {
            "$ref": "#/components/schemas/LstItemCharacteristicEntity"
          },
          "dp_itemGalery": {
            "$ref": "#/components/schemas/LstItemGaleryEntity"
          }
        },
        "required": [
          "dp_id",
          "dp_name",
          "dp_model",
          "dp_cost",
          "dp_photoUrl",
          "dp_itemCategoryId",
          "dp_seoKeywords",
          "dp_seoDescription",
          "dp_itemCharecteristics",
          "dp_itemGalery"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "Розетка тройная встравая с заземлением"
          },
          "dp_model": {
            "type": "string",
            "example": "dp-10012"
          },
          "dp_cost": {
            "type": "number",
            "example": 10.61
          },
          "dp_photoUrl": {
            "type": "string",
            "example": "https://ooodepa.github.io/ooodepa_imgs/products/de-pa-electric/accessories/10012.png"
          },
          "dp_itemCategoryId": {
            "type": "number",
            "example": 1
          },
          "dp_seoKeywords": {
            "type": "string",
            "example": ""
          },
          "dp_seoDescription": {
            "type": "string",
            "example": ""
          },
          "dp_itemCharecteristics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LstItemCharacteristicNoIdDto"
            }
          },
          "dp_itemGalery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LstItemGaleryNoIdDto"
            }
          }
        },
        "required": [
          "dp_name",
          "dp_model",
          "dp_cost",
          "dp_photoUrl",
          "dp_itemCategoryId",
          "dp_seoKeywords",
          "dp_seoDescription",
          "dp_itemCharecteristics",
          "dp_itemGalery"
        ]
      },
...
      "UpdateBulkItemDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemWithIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
...
    }
  }
}